Add files via upload
##Hardware/Software Requirements :      
1 Processor – Intel Core 2 duo/i3/i5/i7 - 64 bit processors are
recommended. The project has been developed in intel i7-64 bit.
2 Operating System – Windows 10: 64-bit
For the purpose of this project we have used Windows 10.
3 Programming Languages :
 * For the backend operations PHP 7 has been used along
with MySQL connectivity via XAMPP.
 * For the frontend Bootstrap 4 has been used
*  For databases, we have used MySQL.
4 Text Editor – Notepad++
Notepad++ is a text editor and source code editor for use with
Microsoft Windows. It supports tabbed editing, which allows
working with multiple open files in a single window. The
project's name comes from the C increment operator. Notepad++
is distributed as free software.
5 Server :
The server has been hosted via Xampp 7.3.10